+++ example-tsp/src/main/java/es/urjc/etsii/grafo/TSP/model/TSPSolution.java
@@ -67,8 +67,7 @@ public class TSPSolution extends Solution<TSPSolution, TSPInstance> {
      *
      * @return current solution score as double
      */
-    @Override
-    public double getScore() {
+    public double getDistance() {
         return this.routeLength;
     }
 
@@ -80,15 +79,10 @@ public class TSPSolution extends Solution<TSPSolution, TSPInstance> {
     }
 
     /**
-     * Recalculate solution score and validate current solution state
-     * You must check that no constraints are broken, and that all costs are valid
-     * The difference between this method and getScore is that we must recalculate the score from scratch,
-     * without using any cache/shortcuts.
-     * DO NOT UPDATE CACHES / MAKE SURE THIS METHOD DOES NOT HAVE SIDE EFFECTS
-     *
+     * Recalculate solution score, sometimes is faster than incrementally calculating it.
+     * This method is also used in the solution validator to verify that the solution is correct.
      * @return current solution score as double
      */
-    @Override
     public double recalculateScore() {
         double distance = 0;
         for (int i = 0; i < this.route.length; i++) {