+++ integration-tests/src/test/java/es/urjc/etsii/morktests/ComplexityAnalysisTest.java
@@ -1,15 +1,15 @@
 package es.urjc.etsii.morktests;
 
-import es.urjc.etsii.grafo.autoconfigtests.model.ACInstance;
-import es.urjc.etsii.grafo.autoconfigtests.model.ACSolution;
-import es.urjc.etsii.grafo.solution.Objective;
+import es.urjc.etsii.grafo.autoconfigtests.Main;
 import es.urjc.etsii.grafo.solver.Mork;
 import es.urjc.etsii.grafo.util.ConcurrencyUtil;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.net.Socket;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.concurrent.TimeUnit;
@@ -28,15 +28,32 @@ public class ComplexityAnalysisTest {
         }
     }
 
+    static void waitPortClosed(int port) throws InterruptedException{
+        // Wait until the webserver is stopped
+        long startCheck = System.nanoTime();
+        boolean closed = false;
+        while (!closed && System.nanoTime() - startCheck < TimeUnit.SECONDS.toNanos(10)) {
+            Thread.sleep(1_000);
+            try(var socket = new Socket("localhost", port)) {
+            } catch (IOException e) {
+                closed = true;
+            }
+        }
+    }
+
+    @AfterAll
+    static void cleanup() throws Exception {
+        waitPortClosed(8080);
+    }
+
     @Test
     void testComplexity() {
-        Objective<?, ACSolution, ACInstance> objective = Objective.ofDefaultMaximize();
         var success = Mork.start(new String[]{
                 "--server.port=0",
                 "--instance-properties",
                 "--instances.path.default=instancesautoconfig/sleepy",
                 "--event.webserver.stopOnExecutionEnd=true",
-        }, objective);
+        }, Main.AC_OBJECTIVE);
         Assertions.assertTrue(success);
         Assertions.assertTrue(Files.exists(propertiesPath));
 
@@ -52,7 +69,7 @@ public class ComplexityAnalysisTest {
                 "--serializers.solution-json.frequency=all",
                 "--serializers.solution-json.folder=timestats",
                 "--event.webserver.stopOnExecutionEnd=true"
-        }, objective);
+        }, Main.AC_OBJECTIVE);
         Assertions.assertTrue(success);
         Assertions.assertTrue(Files.exists(Path.of("timestats")));
         // TODO call python script to analyze results automatically