+++ template/src/main/java/es/urjc/etsii/grafo/__RNAME__/model/__RNAME__Config.java
@@ -4,6 +4,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
+// Note that 'custom' is an example prefix, and can be replaced by whatever you want
 @ConfigurationProperties(prefix = "custom")
 public class __RNAME__Config {
     // Define any configuration property to automatically set from any source,
@@ -12,7 +13,8 @@ public class __RNAME__Config {
 
     /**
      * Value from: custom.my-property, for example from the application.yml file,
-     * could be overridden as a command line parameter: --custom.my-property=value
+     * can also be overridden as a command line parameter: --custom.my-property=value
+     * Upper case letters are converted to lower case and separated by hyphens automatically if necessary
      */
     private String myProperty;