+++ template/src/main/java/es/urjc/etsii/grafo/__RNAME__/model/__RNAME__SolutionValidator.java
@@ -18,13 +18,22 @@ public class __RNAME__SolutionValidator extends SolutionValidator<__RNAME__Solut
      */
     @Override
     public ValidationResult validate(__RNAME__Solution solution) {
-        // Example check:
+        // You should implement this method to check that the solution is valid, without using any kind of existing caches or scores.
+        // For example, you can recalculate solution score and check if it matches the score stored in the solution.
+
+        // Example checks:
 
 //        if(solution.getAssignedElements() > 10){
 //            return ValidationResult.fail("Cannot have more than 10 assigned elements");
 //        }
-//        if(solution.myArray.length == 0){
-//            return ValidationResult.fail("Empty array data, should have at least 5 elements assigned");
+
+//        double recalculateScore = {......};
+//        if(solution.getScore() != recalculateScore){
+//            return ValidationResult.fail("Score mismatch, expected: " + recalculateScore + ", got: " + solution.getScore());
+//        }
+
+//        if(!solution.unassignedClients.isEmpty()){
+//            return ValidationResult.fail("Invalid solution, all clients should be assigned. Remaining clients: " + solution.unassignedClients);
 //        }
 
         return ValidationResult.ok();